[jk] Move logic to DestinationBlock
