+++ mage_ai/data_preparation/models/block/integration/__init__.py
@@ -314,6 +314,30 @@ class SourceBlock(IntegrationBlock):
 
 
 class DestinationBlock(IntegrationBlock):
+    def update(self, data, update_state=False):
+        if update_state:
+            from mage_ai.data_preparation.models.pipelines.integration_pipeline \
+                import IntegrationPipeline
+            from mage_integrations.destinations.utils \
+                import update_destination_state_bookmarks
+
+            integration_pipeline = IntegrationPipeline(self.pipeline.uuid)
+            tap_stream_id = data.get('tap_stream_id')
+            destination_table = data.get('destination_table')
+            bookmark_values = data.get('bookmark_values', {})
+            if tap_stream_id and destination_table:
+                destination_state_file_path = integration_pipeline.destination_state_file_path(
+                    destination_table=destination_table,
+                    stream=tap_stream_id,
+                )
+                update_destination_state_bookmarks(
+                    destination_state_file_path,
+                    tap_stream_id,
+                    bookmark_values=bookmark_values
+                )
+
+        return super().update(data)
+
     def output_variables(self, execution_partition: str = None) -> List[str]:
         return []
